+++ src/Rule/GuessableData.php
@@ -154,10 +154,10 @@ final class GuessableData implements Rule
     }
 
     /**
-     * @param Traversable<string> $words Words to filter.
+     * @param iterable<string> $words Words to filter.
      * @return Traversable<string> Unique words.
      */
-    private function getUniqueWords(Traversable $words): Traversable
+    private function getUniqueWords(iterable $words): Traversable
     {
         $checked = [];
         foreach ($words as $word) {
@@ -179,7 +179,7 @@ final class GuessableData implements Rule
         yield $word;
 
         foreach ($this->wordFormatters as $wordFormatter) {
-            foreach ($wordFormatter->apply($word) as $formatted) {
+            foreach ($wordFormatter->apply([$word]) as $formatted) {
                 yield $formatted;
             }
         }