+++ src/WordList/Pspell.php
@@ -122,7 +122,7 @@ final class Pspell implements WordList
         yield $word;
 
         foreach ($this->wordFormatters as $wordFormatter) {
-            foreach ($wordFormatter->apply($word) as $formatted) {
+            foreach ($wordFormatter->apply([$word]) as $formatted) {
                 yield $formatted;
             }
         }